refactor(crypto): PR #633 review r4 — extract axesForAssetClass util, unify Overall copy, doc/test polish
